Move events and variables out of Join class
