+++ src/main/kotlin/one/oktw/galaxy/Main.kt
@@ -28,8 +28,8 @@ import net.minecraft.server.dedicated.MinecraftDedicatedServer
 import net.minecraft.util.Identifier
 import one.oktw.galaxy.chat.Exchange
 import one.oktw.galaxy.command.CommandRegister
-import one.oktw.galaxy.command.commands.Join
 import one.oktw.galaxy.event.EventManager
+import one.oktw.galaxy.player.PlayerControl
 import one.oktw.galaxy.resourcepack.ResourcePack
 
 @Suppress("unused")
@@ -39,6 +39,8 @@ class Main : ModInitializer {
     lateinit var eventManager: EventManager
         private set
     lateinit var commandRegister: CommandRegister
+        private set
+    lateinit var playerControl: PlayerControl
         private set
 
     companion object {
@@ -51,6 +53,7 @@ class Main : ModInitializer {
         server = FabricLoader.getInstance().gameInstance as MinecraftDedicatedServer
         eventManager = EventManager(server)
         commandRegister = CommandRegister()
+        playerControl = PlayerControl.new()
         Exchange(eventManager)
         main = this
         val resourcePackUrl: String? = System.getenv("resourcePack")
@@ -62,7 +65,5 @@ class Main : ModInitializer {
                 }
             }
         }
-        //Events
-        Join.registerEvent()
     }
 }
